examples/docbuilder/python-docx-service/requirements.txt
flask
python-docx
pillow
requests
